processJoystick now return float

+++ RC_Controller/ControlReader.cpp
@@ -22,13 +22,13 @@ void ControlReader::readControlSpecs(ControlSpecs &controlSpecs) const {
 }
 
 //map the ADC reading to [-1, 1]
-const double ControlReader::processJoystick(int pinId, int midPoint) const {
+const float ControlReader::processJoystick(int pinId, int midPoint) const {
     int rawValue = analogRead(pinId);
 
     if (rawValue < midPoint - 1) {
-        return (midPoint - rawValue) / (double)midPoint + 1;
+        return (midPoint - rawValue) / (float)midPoint + 1;
     } else if (rawValue > midPoint + 1) {
-        return -1 * (rawValue - midPoint) / (double)(1023 - midPoint) + 1;
+        return -1 * (rawValue - midPoint) / (float)(1023 - midPoint) + 1;
     } else {
         return 1;
     }
